Final Path Grayboxing
Took some time to prototype the third, final pathway. We have a tower for the player to ascend, so I was thinking of having some big, interior tower for the player to navigate upward, using broken pipes, catwalks, and grapple points. The top location is where the supposed "boss fight" is taking place, so we would need to cover the exit during that. Or, we would have a completely different way for the player to leave. Not entirely sure yet.
Either way, those are the only major changes this commit.
It's particularly important that, if we do have a "player ascent", I would like the player to see the exit at the top when they first enter the room. Makes it pretty clear what their goal is.