Newer Server Room
Started stitching in the newer server room, which is much wider/longer. I figured I needed to remove some of the geometry from the original and now there is "plenty" of room to see the RAM stick that has to be plugged in. Plus the falling garbage looks a bit more huge and dramatic.
Most of the paths are not stitched in yet. Working on this.