Work on Etherpath.
I'm rethinking some of the graybox designs. The big L room really bothers me so I'm thinking about completely ripping that out and showing the puzzle catwalk another way.
Worked a bit on expanding the original path Natalie created. I'm finding that second story catwalks are a good way to build that spatial awareness, and brings more value to a single room when you have to traverse it a couple separate times.
Etherstitch path and the big catwalk room has been hidden for now. Puzzle room might also be moved, which will require rethinking how the ethernet cord needs to be placed on that path. It doesn't have to always be visible, we can have it snaking through the ceiling or something similar. Might be nice to have the player follow it a bit after it's activated.