Base of the new map management + Map making systems.
I had copied it from one I was making originally and hadnt finished. This has support for steam workshop and discord SDK both of which arent enabled. You can launch the game frmo this tools window, build maps and do other stuff. Will work on it more in the future